tests/submission/Ms2ger/the-elements-of-html/embedded-content/the-video-element/video-tabindex.html

<!DOCTYPE html>
<title>tabindex on video elements</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#video">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="test">
<video></video>
</div>
<script>
test(function() {
  var el = document.getElementById("test").getElementsByTagName("video")[0];
  assert_equals(el.hasAttribute("tabindex"), false);
}, "Attributes shouldn't magically appear");
</script>
